+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# v4.3.0 (Wed Feb 26 2025)
+
+#### 🚀 Enhancement
+
+- feat: add includedTypes and excludedTypes options for selective mock generation [#179](https://github.com/ardeois/graphql-codegen-typescript-mock-data/pull/179) ([@huseyinbuyukdere](https://github.com/huseyinbuyukdere) huseyin_buyukdere@epam.com)
+
+#### Authors: 2
+
+- Hüseyin Büyükdere ([@huseyinbuyukdere](https://github.com/huseyinbuyukdere))
+- huseyinbuyukdere (huseyin_buyukdere@epam.com)
+
+---
+
+# v4.2.0 (Thu Feb 06 2025)
+
+#### 🚀 Enhancement
+
+- Add support for oneOf directive [#180](https://github.com/ardeois/graphql-codegen-typescript-mock-data/pull/180) ([@thomasobrien99](https://github.com/thomasobrien99))
+
+#### Authors: 1
+
+- [@thomasobrien99](https://github.com/thomasobrien99)
+
+---
+
 # v4.1.0 (Fri Sep 20 2024)
 
 #### 🚀 Enhancement